test: scope the integration-test chromium download to web_search
The `_install_playwright` fixture was session-scoped and autouse, so every
integration matrix job downloaded 285 MB of browser. Each job is a fresh
container, so the download never persisted. Only `tests/web_search` uses a
browser.

Move the fixture into `tests/web_search/conftest.py` and add `--only-shell`.
OnyxWebCrawler launches headless with no channel, so the 106 MB headless
shell is enough; the 179 MB full build is not needed.

Drop the `DISABLE_VECTOR_DB` and `shutil.which` guards. The onyx-lite lane
only runs `tests/no_vectordb`, and playwright is a pinned backend dependency,
so both guards are now dead. Delete the no-op overrides in the craft suites
for the same reason.

Result: 50 of 51 pull request jobs download nothing.
